Skip refresh error for data source with depends_on
If a data source has an error during refresh, but has a dependency
declared via depends_on (either directly in the config, or implicitly
through a parent module setting depends_on), log the error and defer
handling of the data source to the plan phase.

If the data source has any depends_on dependencies, then there should
not be any refresh-time requirement for this data source, e.g. provider
configuration, otherwise the initial application of the configuration
would have failed.